# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/lr3/exchange_rate/daily":{"get":{"operationId":"BankController_getСurrencyByDate","parameters":[],"responses":{"200":{"description":""}},"tags":["lr3"]}},"/lr3/exchange_rate/year":{"get":{"operationId":"BankController_getCurrenceByYear","parameters":[{"name":"date_from","required":true,"in":"query","schema":{"type":"string"}},{"name":"date_to","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":"Exchange Rate for date interval","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ExchangeRate"}}}}}},"tags":["lr3"]}},"/lr3/exchange_rate/count":{"get":{"operationId":"BankController_getCountByYear","parameters":[{"name":"date_from","required":true,"in":"query","schema":{"type":"string"}},{"name":"date_to","required":true,"in":"query","schema":{"type":"string"}},{"name":"currency","required":true,"in":"query","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"201":{"description":"Exchange count rate for date interval","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BankReport"}}}}}},"tags":["lr3"]}},"/lr2/mail/login":{"post":{"operationId":"MailController_signIn","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/signInDto"}}}},"responses":{"201":{"description":""}},"tags":["lr2"]}},"/lr2/mail/sendMessage":{"post":{"operationId":"MailController_sendMessage","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/sendMessageDto"}}}},"responses":{"201":{"description":""}},"tags":["lr2"],"security":[{"bearer":[]}]}},"/lr2/mail/getMessage":{"get":{"operationId":"MailController_getMessage","parameters":[],"responses":{"200":{"description":""}},"tags":["lr2"],"security":[{"bearer":[]}]}},"/lr1/feedback/{id}":{"get":{"operationId":"FeedbackController_getFeedback","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"201":{"description":"Модель обратной связи","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Feedback"}}}}}},"tags":["lr1"]}},"/lr1/feedback":{"post":{"operationId":"FeedbackController_createFeedback","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createFeedbackBody"}}}},"responses":{"201":{"description":""}},"tags":["lr1"]}},"/lr1/feedback/user":{"post":{"operationId":"UserController_createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createUserBody"}}}},"responses":{"201":{"description":""}},"tags":["lr1"]}},"/lr1/feedback/tag":{"post":{"operationId":"TagController_createTag","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createTagBody"}}}},"responses":{"201":{"description":""}},"tags":["lr1"]}},"/lr1/feedback/market/{id}":{"get":{"operationId":"MarketController_getMarket","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["lr1"]}},"/lr1/feedback/market":{"post":{"operationId":"MarketController_createMarket","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/createMarketBody"}}}},"responses":{"201":{"description":""}},"tags":["lr1"]}}},"info":{"title":"lr example","description":"The example API description","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"ExchangeRate":{"type":"object","properties":{"date":{"format":"date-time","type":"string"},"currency":{"type":"string"},"value":{"type":"number"}},"required":["date","currency","value"]},"CurrencyCount":{"type":"object","properties":{"min":{"type":"number"},"max":{"type":"number"},"avg":{"type":"number"}},"required":["min","max","avg"]},"BankReport":{"type":"object","properties":{"currency":{"type":"string"},"count":{"$ref":"#/components/schemas/CurrencyCount"}},"required":["currency","count"]},"signInDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"sendMessageDto":{"type":"object","properties":{"to":{"type":"string"},"subject":{"type":"string"},"text":{"type":"string"}},"required":["to","subject","text"]},"User":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"email":{"type":"string"}},"required":["id","name","email"]},"Feedback":{"type":"object","properties":{"id":{"type":"number"},"user":{"$ref":"#/components/schemas/User"},"text":{"type":"string"}},"required":["id","user","text"]},"createFeedbackBody":{"type":"object","properties":{"user_id":{"type":"number"},"market_id":{"type":"number"},"tags_ids":{"type":"array","items":{"type":"number"}},"text":{"type":"string"}},"required":["user_id","market_id","tags_ids","text"]},"createUserBody":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"}},"required":["name","email"]},"createTagBody":{"type":"object","properties":{"name":{"type":"string"},"color":{"type":"string"}},"required":["name","color"]},"createMarketBody":{"type":"object","properties":{"name":{"type":"string"},"address":{"type":"string"}},"required":["name","address"]}}}}